requirements.txt
flask
gunicorn
networkx
python-dotenv
google-genai
pytest
pytest-cov
pre-commit
ruff